Add template specialisation for integer uniform distribution in poprand
Summary:
The integer distribution is generated by using multiprecision unsigned
integer arithmetic

Test Plan: Unit test added

Reviewers: Eddy

Reviewed By: Eddy

Subscribers: davidn

Differential Revision: https://phabricator.sourcevertex.net/D2262

+++ include/poprand/RandomGen.hpp
@@ -38,35 +38,47 @@ enum RandomGenMode {
   NOT_REPEATABLE
 };
 
-/// Uniform distribution in [minVal, maxVal] with maxVal > minVal
-/// The mode determines whether the numbers generated are
+/// Uniform distribution in a given interval with maxVal >
+/// minVal The mode determines whether the numbers generated are
 /// repeatable across systems
-void uniform(poplar::Graph &graph, poplar::Tensor &A, float minVal,
-             float maxVal, uint64_t seed, RandomGenMode mode,
+///
+/// The tensor A may be of type "float", "half" or "int".
+/// For "float" and "half", the uniform number generator
+/// generates data with the uniform distribution in the interval
+/// [minVal, maxVal]
+/// For "int", data is generated in interval [minVal, maxVal)
+/// with uniform probability if maxVal - minVal is a power of 2.
+/// Otherwise there will be a small bias in the probability
+/// generated with the bias directly proportional to the ratio
+/// maxVal-minVal / 2^32.
+void uniform(poplar::Graph &graph, poplar::Tensor &A, double minVal,
+             double maxVal, uint64_t seed, RandomGenMode mode,
              poplar::program::Sequence &prog,
              const std::string &debugPrefix = "");
-void uniform(poplar::Graph &graph, poplar::Tensor &A, float minVal,
-             float maxVal, RandomGenMode mode, poplar::program::Sequence &prog,
+void uniform(poplar::Graph &graph, poplar::Tensor &A, double minVal,
+             double maxVal, RandomGenMode mode, poplar::program::Sequence &prog,
              const std::string &debugPrefix = "");
 
 /// Bernoulli with probablility of 1 = "prob"
 /// The mode determines whether the numbers generated are
 /// repeatable across systems
-void bernoulli(poplar::Graph &graph, poplar::Tensor &A, float prob,
+/// Tensor types supported are "float", "half" and "int"
+void bernoulli(poplar::Graph &graph, poplar::Tensor &A, double prob,
                uint64_t seed, RandomGenMode mode,
                poplar::program::Sequence &prog,
                const std::string &debugPrefix = "");
-void bernoulli(poplar::Graph &graph, poplar::Tensor &A, float prob,
+void bernoulli(poplar::Graph &graph, poplar::Tensor &A, double prob,
                RandomGenMode mode, poplar::program::Sequence &prog,
                const std::string &debugPrefix = "");
 
 /// Normal distribution with given mean and standard deviation
 /// The mode determines whether the numbers generated are
 /// repeatable across systems
-void normal(poplar::Graph &graph, poplar::Tensor &A, float mean, float stdDev,
+/// The tensor A may be of type "half" and "float"
+void normal(poplar::Graph &graph, poplar::Tensor &A, double mean, double stdDev,
             uint64_t seed, RandomGenMode mode, poplar::program::Sequence &prog,
             const std::string &debugPrefix = "");
-void normal(poplar::Graph &graph, poplar::Tensor &A, float mean, float stdDev,
+void normal(poplar::Graph &graph, poplar::Tensor &A, double mean, double stdDev,
             RandomGenMode mode, poplar::program::Sequence &prog,
             const std::string &debugPrefix = "");
 
@@ -77,12 +89,13 @@ void normal(poplar::Graph &graph, poplar::Tensor &A, float mean, float stdDev,
 ///   (mean - alpha * stdDev) and (mean + alpha * stdDev)
 /// The mode determines whether the numbers generated are
 /// repeatable across systems
-void truncatedNormal(poplar::Graph &graph, poplar::Tensor &A, float mean,
-                     float stdDev, float alpha, uint64_t seed,
+/// The tensor A may be of type "half" and "float"
+void truncatedNormal(poplar::Graph &graph, poplar::Tensor &A, double mean,
+                     double stdDev, double alpha, uint64_t seed,
                      RandomGenMode mode, poplar::program::Sequence &prog,
                      const std::string &debugPrefix = "");
-void truncatedNormal(poplar::Graph &graph, poplar::Tensor &A, float mean,
-                     float stdDev, float alpha, RandomGenMode mode,
+void truncatedNormal(poplar::Graph &graph, poplar::Tensor &A, double mean,
+                     double stdDev, double alpha, RandomGenMode mode,
                      poplar::program::Sequence &prog,
                      const std::string &debugPrefix = "");
